1. Structure tests 2. Add test that tests get_response function from rss_reader module with bad url
